Rename index.ts files to be more descriptive.
